+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
+- Fixed a bug where the invoice amount wasn’t formatted correctly for zero-decimal currencies. ([#96](https://github.com/craftcms/stripe/issues/96))
 - `craft\stripe\elements\Product::getPrices()` now has an optional `$criteria` argument. ([#97](https://github.com/craftcms/stripe/issues/97))
+
 
 ## 1.6.0 - 2025-06-25